fix(model): refactor ip router model
